리뷰 내용 반영
- System.out.println 제거(로거를 사용할 수 있도록 할 것)
- config 클래스 분리
